% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaz.R
\name{doePractice}
\alias{doePractice}
\title{Interfaz de usuario: Generador de Ejercicios de Diseño de Experimentos}
\usage{
doePractice()
}
\description{
Ver 0.2.0 - Incorporando temas del Parcial 2
Ver 0.1.0 - Versión inicial.
}
