% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examen.R
\name{genExamen}
\alias{genExamen}
\title{Generador de Examen}
\usage{
genExamen(parcial = NULL, format = "HTML", lang = "Esp")
}
\arguments{
\item{parcial}{1 o 2, según el parcial que se vaya a utilizar}

\item{format}{HTML o LaTeX, formato en que se generara el archivo de salida del examen}

\item{lang}{Esp o Eng. Idioma en que se armara el archivo de salida}
}
\value{
Genera una tabla con los enunciados del examen, un archivo de salida de texto (HTML o LaTeX), y actualiza el directorio
}
\description{
Version 2.8.1 - Cambios menores para Quality of Life. Se añaden pausas en la interfaz para el usuario.
Version 2.8.0 - Se incorporan el framework para trabajar con los temas B y K
Version 2.7.5 - Correccion del idioma y los problemas con el orden de los enunciados
Version 2.7.4 - Se separa el ensamblador para facilitar el trabajar con la tabla y poder regenerar el formato sin tener que regenerar la tabla
Version 2.7.3 - Correccion para que no se repitan los problemas de ANOVA.
Version 2.7.2 - Correcciones menores de formato para cli
Version 2.7.1 - Se cambian los ´on.exit(sink())´ y ´sink()´ por  ´closeAllConnections()´
Version 2.7 - usando el paquete cli para explicarle al usuario todo lo que se esta haciendo en la función (eg. llamar a las tablas, etc.)
Version 2.6.9 - ajuste en el query y argumentos para acomodar a la interfaz doeExam()
Version 2.6.8 - usar los paquetes here y folders para los paths relativos
Version 2.6.7 - Se llama a la generacion del reporte de resultados inmediatamente que se genera el examen
Version 2.6.6 - Ajuste para funcionar dentro del paquete, tablas externas se guardan como internas y se añaden dependencia de paquetes
Version 2.6.5
}
