% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{metadataDB}
\alias{metadataDB}
\title{Generar metadatos genericos para la base de datos}
\usage{
metadataDB(Data, Tema, Cont)
}
\arguments{
\item{Data}{Tabla de datos a la cual generar los metadatos}

\item{Tema}{Tema de la tabla: MM - Media Muestral, PH - Prueba de Hipotesis, A1 - Anova, etc}

\item{Cont}{Tipo de tabla: Datos muestrales, Preguntas, Examen, etc.}
}
\description{
V2.1 -
V2 - Ahora la función loopea si detecta que se generaron id's repetidos en el Directorio
}
