% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaz.R
\name{poblarDirectorio}
\alias{poblarDirectorio}
\title{Crear tablas para un Directorio.}
\usage{
poblarDirectorio(tema = NULL, aislado = FALSE)
}
\arguments{
\item{tema}{Numero del tema (1-5), donde 1 = Media Muestral, 2 = Prueba de Hipotesis, 3 = Anova 1Factor, 4 = Anova 1F+Bloque, 5 = Diseño 2^k}

\item{aislado}{Flag para saber si se genera un Directorio temporal o no.}
}
\description{
Ver 2.1.0 - Implementando tameas 4 y 5
Ver 2.0.0 - Ajuste mayor: Ahora se genera tabla tema por tema. Se debe llamar a la función otra vez si se desea un tema distinto.
Ver 1.0.0 - Inicial, se generan todas las tablas de temas del parcial
}
