% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generadores.R
\name{question1FBlock}
\alias{question1FBlock}
\title{Anova 1F + Bloque: Generador de Preguntas}
\usage{
question1FBlock(DB, Q, format = "HTML", lang = "Esp", silent = T)
}
\arguments{
\item{DB}{Tabla de datos muestrales a usar para la generación de preguntas}

\item{Q}{Numero de pregunas a generar}

\item{format}{Formato para los enunciados de preguntas. "HTML" o "LaTeX".}

\item{lang}{Eng o Esp. Idiomas para los enunciados}

\item{silent}{Si \code{TRUE}, enonces no se abre el archivo de texto generado}
}
\value{
Una tabla de datos, se guarda en un archivo, y se puede asignar a un objeto
}
\description{
Ver 2.2.1 - Ajustes menores
Ver 2.2 - Integración al Directorio
Ver 2.1 - Integracion del idioma con tablas externas
Ver 2.0 - Solo HTML
}
