% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generadores.R
\name{question2k}
\alias{question2k}
\title{Diseño 2^k: Generador de preguntas}
\usage{
question2k(DB, Q, format = "HTML", lang = "Esp", silent = T)
}
\arguments{
\item{DB}{Tabla de datos muestrales a usar para la generación de preguntas}

\item{Q}{Numero de pregunas a generar}

\item{format}{Formato para los enunciados de preguntas. "HTML" o "LaTeX".}

\item{lang}{Eng o Esp. Idiomas para los enunciados}

\item{silent}{Si \code{TRUE}, enonces no se abre el archivo de texto generado}
}
\value{
Una tabla de datos, se guarda en un archivo, y se puede asignar a un objeto
}
\description{
Ver 2.2.2 - Se incluye la tabla de anova en el enunciado - se debe eliminar todo menos el SST
Ver 2.2.1 - Cambios menores
Ver 2.2.0 - Integración del directorio
Ver 2.1.0 - Incorporacion de idioma
Ver 2.0.0 - Inicial
}
